(newapp) Remove `app/auth/auth-utils.ts` by moving code into the login mutation
