--- packages/generator/templates/app/app/auth/auth-utils.ts
@@ -1,18 +1,0 @@
-import { SecurePassword, AuthenticationError } from "blitz"
-import db from "db"
-
-export const authenticateUser = async (email: string, password: string) => {
-  const user = await db.user.findFirst({ where: { email } })
-  if (!user) throw new AuthenticationError()
-
-  const result = await SecurePassword.verify(user.hashedPassword, password)
-
-  if (result === SecurePassword.VALID_NEEDS_REHASH) {
-    // Upgrade hashed password with a more secure hash
-    const improvedHash = await SecurePassword.hash(password)
-    await db.user.update({ where: { id: user.id }, data: { hashedPassword: improvedHash } })
-  }
-
-  const { hashedPassword, ...rest } = user
-  return rest
-}
